+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿# Changelog
+# Changelog
 
 All notable changes to GrossGeo.SDK.Stub will be documented in this file.
 
@@ -6,6 +6,15 @@ Format: [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 Versioning: [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- TestProduct.Installer (DistributionType.Installer)
+- TestProduct.PluginDll (DistributionType.PluginDll)
+
+### Changed
+- All 8 samples: manifests synced with code
+- All README.md updated
+
 
 ## [1.0.0] - 2026-03-02